[cheriabi] Use capability type for Null/Crap pools.
Ensure that the Null and Crap pools types are capable of storing
capabilities and that their sizes are rounded to the correctly
alignment.

+++ src/hb-null.hh
@@ -104,7 +104,7 @@ using hb_min_size = _hb_min_size<T, void>;
  */
 
 extern HB_INTERNAL
-uint64_t const _hb_NullPool[(HB_NULL_POOL_SIZE + sizeof (uint64_t) - 1) / sizeof (uint64_t)];
+uintptr_t const _hb_NullPool[(HB_NULL_POOL_SIZE + sizeof (uintptr_t) - 1) / sizeof (uintptr_t)];
 
 /* Generic nul-content Null objects. */
 template <typename Type>
@@ -169,7 +169,7 @@ struct NullHelper
  * causing bad memory access. So, races there are not actually introducing incorrectness
  * in the code. Has ~12kb binary size overhead to have it, also clang build fails with it. */
 extern HB_INTERNAL
-/*thread_local*/ uint64_t _hb_CrapPool[(HB_NULL_POOL_SIZE + sizeof (uint64_t) - 1) / sizeof (uint64_t)];
+/*thread_local*/ uintptr_t _hb_CrapPool[(HB_NULL_POOL_SIZE + sizeof (uintptr_t) - 1) / sizeof (uintptr_t)];
 
 /* CRAP pool: Common Region for Access Protection. */
 template <typename Type>
