+++ src/hb-static.cc
@@ -43,8 +43,8 @@
 #ifndef HB_NO_VISIBILITY
 #include "hb-ot-name-language-static.hh"
 
-uint64_t const _hb_NullPool[(HB_NULL_POOL_SIZE + sizeof (uint64_t) - 1) / sizeof (uint64_t)] = {};
-/*thread_local*/ uint64_t _hb_CrapPool[(HB_NULL_POOL_SIZE + sizeof (uint64_t) - 1) / sizeof (uint64_t)] = {};
+uintptr_t const _hb_NullPool[(HB_NULL_POOL_SIZE + sizeof (uintptr_t) - 1) / sizeof (uintptr_t)] = {};
+/*thread_local*/ uintptr_t _hb_CrapPool[(HB_NULL_POOL_SIZE + sizeof (uintptr_t) - 1) / sizeof (uintptr_t)] = {};
 
 DEFINE_NULL_NAMESPACE_BYTES (OT, Index) =  {0xFF,0xFF};
 DEFINE_NULL_NAMESPACE_BYTES (OT, VarIdx) =  {0xFF,0xFF,0xFF,0xFF};
